requirements.txt
pymupdf
pillow
pytesseract
python-dotenv
langchain
faiss-cpu
